eval_openpi.py
"""
Usage:
python eval.py --checkpoint data/image/pusht/diffusion_policy_cnn/train_0/checkpoints/latest.ckpt -o data/pusht_eval_output
"""

import sys

from diffusion_policy.env_runner.robomimic_openpi_runner import RobomimicOpenPIRunner

# use line-buffering for both stdout and stderr
sys.stdout = open(sys.stdout.fileno(), mode="w", buffering=1)
sys.stderr = open(sys.stderr.fileno(), mode="w", buffering=1)

import os
import pathlib
import click
import hydra
import torch
import dill
import wandb
import json
from diffusion_policy.workspace.base_workspace import BaseWorkspace
from omegaconf import OmegaConf

@hydra.main(
    version_base=None,
    config_path=str(
        pathlib.Path(__file__).parent.joinpath("diffusion_policy", "config", "eval"),
    ),
    config_name="openpi_mug",  # 指定默认配置文件
)
# def main(output_dir, task, max_steps, runner, if_mask):
def main(cfg: OmegaConf):
    OmegaConf.resolve(cfg)
    output_dir = cfg.env_runner.output_dir
    
    if os.path.exists(output_dir):
        click.confirm(
            f"Output path {output_dir} already exists! Overwrite?", abort=True
        )
    pathlib.Path(output_dir).mkdir(parents=True, exist_ok=True)

    env_runner: RobomimicOpenPIRunner = hydra.utils.instantiate(cfg.env_runner, output_dir=output_dir)
    runner_log = env_runner.run()

    # dump log to json
    json_log = dict()
    for key, value in runner_log.items():
        if isinstance(value, wandb.sdk.data_types.video.Video):
            json_log[key] = value._path
        else:
            json_log[key] = value
    out_path = os.path.join(output_dir, "eval_log.json")
    json.dump(json_log, open(out_path, "w"), indent=2, sort_keys=True)


if __name__ == "__main__":
    main()
