requirements.txt
# Python dependencies converted from conda_environment2.yaml
# Core Python packages
# python==3.10
# pip==22.2.2

# PyTorch ecosystem
# torch==1.12.1
# torchvision==0.13.1
pytorch3d==0.7.0

# Scientific computing
numpy==1.23.3
scipy==1.9.1
numba==0.56.4

# Computer vision and image processing
opencv-python
scikit-image==0.19.3
scikit-video==1.1.11
imageio==2.22.0
imageio-ffmpeg==0.4.7
imagecodecs==2022.9.26

# Data handling
h5py==3.7.0
zarr==2.12.0
numcodecs==0.10.2

# Machine learning and AI
hydra-core==1.2.0
einops==0.4.1
wandb==0.13.3
tensorboard==2.10.1
tensorboardx==2.5.1
accelerate==0.13.2
datasets==2.6.1
diffusers==0.11.1

# Robotics and simulation
# gym==0.21.0
pymunk==6.2.1
pygame==2.1.2
pybullet-svl==3.1.6.4
# robosuite @ https://github.com/cheng-chi/robosuite/archive/277ab9588ad7a4f4b55cf75508b44aa67ec171f0.tar.gz
# robomimic==0.2.0
dm-control==1.0.9

# Utilities
tqdm==4.64.1
dill==0.3.5.1
cffi==1.15.1
ipykernel==6.16
matplotlib==3.6.1
threadpoolctl==3.1.0
shapely==1.8.4
cython==0.29.32
termcolor==2.0.1
psutil==5.9.2
click==8.0.4
boto3==1.24.96
av==10.0.0
cmake==3.24.3

# Video processing
pytorchvideo==0.1.5

# R3M for representation learning
r3m @ https://github.com/facebookresearch/r3m/archive/b2334e726887fa0206962d7984c69c5fb09cceab.tar.gz

# Ray for distributed computing
ray[default,tune]==2.2.0

# Free MuJoCo (requires system dependencies: libosmesa6-dev libgl1-mesa-glx libglfw3 patchelf)
free-mujoco-py==2.1.6

# Note: Some packages like cudatoolkit, llvm-openmp are system-level dependencies
# and should be installed separately via system package manager 